Add hash tag support for jarvis
