+++ server/slack/commands.js
@@ -1,0 +1,10 @@
+// All commands are case insensitive
+const Commands = {
+  TIL: ["#til", "til", "#todayilearn"],
+  HELP: ["#help"],
+  CREATE_HANGOUT: ["#createhangout"],
+  LIST_HANGOUTS: ["#listhangouts", "#listhangout"],
+  HOW_ARE_YOU: ["#howareyou", "#howareyou?"]
+};
+
+export default Commands;
